drivers: cpuidle: lpm-levels: Support s2idle
If we support suspend-to-idle we can avoid the overhead in
hotplugging out CPUs and tearing things down during suspend. Most
things are already in place to support this so we can pretty much
re-use the platform suspend ops to enter the freeze state for
suspending with cpuidle. We need to move the clk printing call
into cluster_configure() so that we only call this when we're
going into suspend on the last CPU in the case of freeze. Putting
it here shares code between both versions of suspend.

Change-Id: Ib062d93a0c122a644af4a729dbb202b88c9a6c56
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>
Signed-off-by: Subhajeet Muhuri <kenny3fcb@gmail.com>